+++ entities/max.cpp
@@ -296,7 +296,7 @@ IMPLEMENT_FUNCTION(Max, function9, 9)
 		if (params->param2 == kTimeInvalid || !getState()->time)
 			break;
 
-		if (params->param1 >= (int)getState()->time) {
+		if (params->param1 >= getState()->time) {
 			if (!getEntities()->hasValidFrame(kEntityMax) || !params->param2) {
 
 				params->param2 = getState()->time;
@@ -304,7 +304,7 @@ IMPLEMENT_FUNCTION(Max, function9, 9)
 					goto setup_functions;
 			}
 
-			if (params->param2 >= (int)getState()->time)
+			if (params->param2 >= getState()->time)
 				break;
 		}
